requirements.txt
fastapi
uvicorn
torch
numpy
opencv-python
Pillow
matplotlib
ultralytics